# UrlShortener/UrlShortener/openapi.yaml
openapi: 3.0.3
info:
  title: URL Shortener API
  version: 1.0.0
  description: Simple RESTful API for shortening URLs.
paths:
  /shorten:
    post:
      operationId: Shorten
      summary: Shorten a URL
      requestBody:
        required: true
        content:
          application/json:
            schema:
              title: UrlBody
              type: object
              additionalProperties: false
              required:
                - fullUrl
              properties:
                fullUrl:
                  type: string
                  example: https://example.com/very/long/url
                customAlias:
                  type: string
                  example: my-custom-alias
      responses:
        '201':
          description: URL successfully shortened
          content:
            application/json:
              schema:
                title: UrlResponse
                type: object
                additionalProperties: false
                properties:
                  shortUrl:
                    type: string
                    example: http://localhost:8080/my-custom-alias
        '400':
          description: Invalid input or alias already taken
...